+++ griptape/tools/base_tool.py
@@ -41,7 +41,7 @@ class BaseTool(ActivityMixin, ABC):
     install_dependencies_on_init: bool = field(default=True, kw_only=True)
     dependencies_install_directory: Optional[str] = field(default=None, kw_only=True)
     verbose: bool = field(default=False, kw_only=True)
-    off_prompt: bool = field(default=True, kw_only=True)
+    off_prompt: bool = field(default=False, kw_only=True)
 
     def __attrs_post_init__(self) -> None:
         if self.install_dependencies_on_init: